:sparkles:cssに移植 & コードを整理
