+++ django_quickdraw/settings.py
@@ -128,3 +128,9 @@ USE_TZ = False
 # https://docs.djangoproject.com/en/1.11/howto/static-files/
 
 STATIC_URL = '/static/'
+
+STATIC_ROOT = os.path.join(BASE_DIR, 'static')
+
+STATICFILES_DIRS = (
+    [os.path.join(BASE_DIR, 'quickdraw', 'static')]
+)
