+++ templates/base.html
@@ -4,6 +4,8 @@
   <meta charset="UTF-8">
   <title>{% block title %}Quick Draw Demo{% endblock %}</title>
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
+  {% block quickdraw_style %}
+  {% endblock %}
 </head>
 <body>
   {% block body %}
